// src/entities/articles/model/types.ts
export type ArticleId = string;
export type ArticlesData = {
    articles: Article[];
    articlesCount: number;
};

export type Article = {
    title: string;
    description: string;
    slug: ArticleId;
    favorited: boolean;
    favoritesCount: number;
    tagList: string[];
    createdAt: string;
    body: string | undefined;
    author: {
        username: string;
        image: string;
        following: boolean;
    };
};